chat middle commit
